chore: PR feedback - move IntersectionObserver mock to root mocks folder so it can be re-used with other packages. Update import path for the Stage component tests.
